+++ modules/apps/portal-search/portal-search-spi/src/main/java/com/liferay/portal/search/spi/model/permission/contributor/SearchPermissionFilterContributor.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.spi.model.permission;
+package com.liferay.portal.search.spi.model.permission.contributor;
 
 import com.liferay.portal.kernel.search.filter.BooleanFilter;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
